backend/requirements-dev.txt
-r requirements.txt
ruff>=0.9
pytest>=8.0
pytest-django>=4.9
pytest-cov>=6.0
